backend/requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
Django==5.1.4
djangorestframework==3.15.2
django-environ==0.11.2

# ── Database ──────────────────────────────────────────────────────────────────
mysqlclient==2.2.4               # C extension — fastest MySQL driver for Django

# ── Auth & Security ───────────────────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-axes==7.0.0               # brute-force login protection

# ── Async tasks ───────────────────────────────────────────────────────────────
celery==5.4.0
django-celery-beat==2.7.0        # DB-backed periodic tasks
django-celery-results==2.5.1     # store task results in DB
redis==5.2.0

# ── API utilities ─────────────────────────────────────────────────────────────
django-filter==24.3              # filter querysets in DRF views
drf-spectacular==0.27.2          # OpenAPI / Swagger docs

# ── Payments ──────────────────────────────────────────────────────────────────
requests==2.32.3                 # M-Pesa Daraja API calls
urllib3==1.26.20                 # pin to version africastalking expects (avoids RequestsDependencyWarning)

# ── Notifications ─────────────────────────────────────────────────────────────
africastalking==1.2.5            # SMS + WhatsApp

# ── Files & PDFs ─────────────────────────────────────────────────────────────
Pillow==11.0.0                   # image handling (maintenance photos)
reportlab==4.2.5                 # PDF receipts & lease documents

# ── Production server ─────────────────────────────────────────────────────────
gunicorn==23.0.0

# ── Dev / Testing ─────────────────────────────────────────────────────────────
pytest-django==4.9.0
pytest==8.3.4
factory-boy==3.3.1               # test data factories
